Improve tests of u-case/with_validation
